feat(components): cards, pills, buttons, scrollshadows
