+++ src/app/(authenticated)/settings/page.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { ScrollShadow } from '@nextui-org/react';
 import React from 'react';
 
 import Box from '@/components/Box';
@@ -36,14 +37,16 @@ const Settings: React.FC = () => {
   return (
     <Box className='h-full flex flex-col'>
       <Header title='Settings'></Header>
-      <div className='flex flex-col px-6 my-6 gap-y-6'>
-        {APP_SETTINGS.map((s, i) => (
-          <div key={i} className='flex items-start justify-between'>
-            <h3 className='font-medium'>{s.label}</h3>
-            <h4 className=''>{s.value}</h4>
-          </div>
-        ))}
-      </div>
+      <ScrollShadow hideScrollBar>
+        <div className='flex flex-col px-6 my-6 gap-y-6'>
+          {APP_SETTINGS.map((s, i) => (
+            <div key={i} className='flex items-start justify-between'>
+              <h3 className='font-medium'>{s.label}</h3>
+              <h4 className=''>{s.value}</h4>
+            </div>
+          ))}
+        </div>
+      </ScrollShadow>
     </Box>
   );
 };
